Move theme detection into v wrapper bash function
Since the theme detection code is at the top of ‘wrappers.sh’ which is
sourced in ‘.bashrc’, even if we use ‘THEME=light’ before invoking ‘v’,
the theme is still dark because it was set in stone at the time the file
was sourced. This commit moves that logic inside ‘v’ wrapper function so
that it is re-evaluated every time and enabling the use to really pick
the dark or light theme at will.