Switch vim theme from gruvbox light so solarized light
Gruvbox is amazing but it is not working well for Jest/JavaScript unit
tests. The colors on spec/test files are the same for `describe`, `it`
and the strings that describe each test. A little strange. Solarized
light is offering me a nicer syntax colors for this case and I'll stick
with it for now.