editorconfig: use tabs and text width for bash scripts
Bash has some features related to HEREDOCS and hard tabs to allow
indeting the source code and using spaces for bash script indentation
end up making one losing that goodie. Example:

  if [[ -z $1 ]] ; then
  	cat <<-EOF
  	  Use the source, Luke.
  	  Read the --help, the info and the man pages.
  	EOF
  fi

The snippet above allowed the ending EOF to be indented with the initial
like of the heredoc. Only spaces are kept in the output.