Add extra key comb for kitty open shell
Note `ctrl+shift+esc` does not work in MacOS for some reason.
`Ctrl-<something>` works with some key combinations, not with others.

As a side note, Neomutt was not responding to `Ctrl+O` to open the
current selected sidebar item, and running

  $ stty discard undef

before running Neomutt solved the issue. Did not seem to work for making
`Ctrl+<things>` work with Kitty.

REFERENCES:
• https://github.com/neomutt/neomutt/issues/1195#issuecomment-388016804