Make helm-rg consider use root dir by default
When search with `helm-rg`, the search path is set to the path of the
currently buffer. This causes the search not to happen from the entire
project, but depending on where that buffer's file is located, on a
subdirectory of the project. Until I learn how to handle this in a more
dynamic way, I'll make `helm-rg` **always** search from the root dir so
I can `rg` the entire project.