emacs: Remove newline from yas snippets
It looks like yasnippet inserted newlines inside some snippets causing
their expansion to also add newlines after the expanded text, which is
especially bad for “inline” stuff, like:

   Try `bt<Tab>
   ` and see.

It causes the line break between ` and `.

Looking for yasnipet issues, it seems it should already set

   mode-require-final-newline
   require-final-newline

both to nil, but it is not happening for some reason. I set them
manually when visiting the snippet files and ran:

  (setq mode-require-final-newline nil)
  (setq require-final-newline nil)

Then I removed the last newline and saved the files...