nvim: Disable diagnostic virtual text
The diagnostics that show by default as virtual text on the right of
code lines containing problems is too noisy, and one can't ready it
anyway as the text is sometimes very long, requiring one to view the
popup window with the more readable message. Therefore, disabling it and
if I really want to see the message I navigate to the diagnostic
message, which opens the popup window.